+++ src/main/java/org/apiphany/io/UInt64.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
 
-import org.apiphany.json.JsonBuilder;
 import org.apiphany.lang.annotation.AsValue;
 
 /**
@@ -109,13 +108,13 @@ public class UInt64 implements ByteSizeable, BinaryRepresentable {
 	}
 
 	/**
-	 * Returns a JSON representation of this object.
+	 * Returns a string representation of this object.
 	 *
-	 * @return JSON string representation
+	 * @return string representation
 	 */
 	@Override
 	public String toString() {
-		return JsonBuilder.toJson(this);
+		return String.valueOf(value);
 	}
 
 	/**